resolved all issues raised in the review
