+++ src/screens/sign-up/types/enums.ts
@@ -1,0 +1,5 @@
+export enum Position {
+    ABSOLUTE = "absolute",
+    RELATIVE = "relative",
+  }
+
