+++ test-report.xml
@@ -1,22 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <testExecutions version="1">
-  <file path="/Users/deca/Documents/Tasks/forked/auth-mobile/src/components/shared/banner/banner.test.tsx">
-  </file>
   <file path="/Users/deca/Documents/Tasks/forked/auth-mobile/src/components/button/button.test.tsx">
   </file>
+  <file path="/Users/deca/Documents/Tasks/forked/auth-mobile/src/components/shared/banner/banner.test.tsx">
+  </file>
   <file path="/Users/deca/Documents/Tasks/forked/auth-mobile/src/screens/sign-up/components/FormFieldsComponent.test.tsx">
-    <testCase name="FormFieldsComponent renders correctly with provided props" duration="7556"/>
-    <testCase name="FormFieldsComponent calls handleChange when input value changes" duration="12"/>
+    <testCase name="FormFieldsComponent renders correctly with provided props" duration="451"/>
+    <testCase name="FormFieldsComponent calls handleChange when input value changes" duration="11"/>
   </file>
   <file path="/Users/deca/Documents/Tasks/forked/auth-mobile/src/screens/sign-up/sign-up-screen.test.tsx">
-    <testCase name="SignUpScreen renders correctly" duration="6771"/>
+    <testCase name="SignUpScreen renders correctly" duration="611"/>
   </file>
   <file path="/Users/deca/Documents/Tasks/forked/auth-mobile/src/screens/sign-up/components/LogoComponent.test.tsx">
-    <testCase name="LogoComponent renders correctly with provided props" duration="6907"/>
-    <testCase name="LogoComponent matches the snapshot" duration="20"/>
+    <testCase name="LogoComponent renders correctly with provided props" duration="923"/>
+    <testCase name="LogoComponent matches the snapshot" duration="6"/>
   </file>
   <file path="/Users/deca/Documents/Tasks/forked/auth-mobile/src/screens/sign-up/components/ButtonComponent.test.tsx">
-    <testCase name="ButtonComponent renders with correct props and styles" duration="6936"/>
-    <testCase name="ButtonComponent calls onPress handler when pressed" duration="185"/>
+    <testCase name="ButtonComponent renders with correct props and styles" duration="843"/>
+    <testCase name="ButtonComponent calls onPress handler when pressed" duration="158"/>
   </file>
 </testExecutions>
